+++ tests/neuro_san/internals/run_context/langchain/token_counting/test_llm_token_callback_handler.py
@@ -15,39 +15,18 @@
 #
 # END COPYRIGHT
 
-from contextlib import contextmanager
-import logging
-
-from time import time
 from typing import Dict
 
-from langchain_core.messages import AIMessage
-from langchain_core.outputs import ChatGeneration
-from langchain_core.outputs import LLMResult
+from unittest import TestCase
+
+import logging
+
 import pytest
 
 from neuro_san.internals.run_context.langchain.token_counting.llm_token_callback_handler import LlmTokenCallbackHandler
-from neuro_san.internals.run_context.langchain.token_counting.llm_token_callback_handler import llm_token_callback_var
 
 
-@contextmanager
-def owning_agent_scope(handler: LlmTokenCallbackHandler):
-    """
-    Simulate the agent scope that owns the LLM call being handled.
-
-    In production, each agent's count_tokens() sets llm_token_callback_var to its
-    own handler, so at event time the var holds the handler of the nearest
-    enclosing agent scope.  The handler consults it to tell its own agent's LLM
-    calls apart from downstream agents' calls it also hears about.
-    """
-    token = llm_token_callback_var.set(handler)
-    try:
-        yield
-    finally:
-        llm_token_callback_var.reset(token)
-
-
-class TestLlmTokenCallbackHandler:
+class TestLlmTokenCallbackHandler(TestCase):
     """Test cases for the LlmTokenCallbackHandler.calculate_token_costs method."""
 
     @pytest.fixture
@@ -169,141 +148,3 @@ class TestLlmTokenCallbackHandler:
         cost = handler.calculate_token_costs("gpt-4", completion_tokens, prompt_tokens)
 
         assert abs(cost - expected_cost) < 0.000001  # Account for floating point precision
-
-
-class TestEmptyResponseTracking:
-    """Test cases for the empty_responses tracking added to LlmTokenCallbackHandler."""
-
-    @pytest.mark.parametrize("message,expected", [
-        (AIMessage(content=""), True),
-        (AIMessage(content="   "), True),
-        (AIMessage(content="here is the answer"), False),
-        (AIMessage(content=[]), True),
-        (AIMessage(content=[" "]), True),
-        (AIMessage(content=[{"type": "text", "text": ""}]), True),
-        (AIMessage(content=[{"type": "text", "text": "   "}]), True),
-        (AIMessage(content=[{"type": "text", "text": "real text"}]), False),
-        (
-            AIMessage(
-                content="",
-                tool_calls=[{"name": "search", "args": {}, "id": "1", "type": "tool_call"}],
-            ),
-            False,
-        ),
-    ])
-    def test_is_empty_response(self, message, expected):
-        """A response is empty only when it carries neither content nor a tool call."""
-        # pylint: disable=protected-access
-        assert LlmTokenCallbackHandler._is_empty_response(message) is expected
-
-    def _make_result(self, content, tool_calls=None) -> LLMResult:
-        """Build an LLMResult wrapping a single AIMessage with usage metadata."""
-        message = AIMessage(
-            content=content,
-            tool_calls=tool_calls or [],
-            usage_metadata={"input_tokens": 10, "output_tokens": 0, "total_tokens": 10},
-            response_metadata={"model_name": "gpt-oss:20B"},
-        )
-        return LLMResult(generations=[[ChatGeneration(message=message)]])
-
-    def _make_handler(self) -> LlmTokenCallbackHandler:
-        handler = LlmTokenCallbackHandler(llm_infos={})
-        handler.provider_class = "ollama"
-        handler.models_token_dict = {"ollama": {}}
-        handler.start_time = time()
-        return handler
-
-    @pytest.mark.asyncio
-    async def test_on_llm_end_counts_empty_response(self):
-        """An empty response counts as a successful request AND an empty one."""
-        handler = self._make_handler()
-        with owning_agent_scope(handler):
-            await handler.on_llm_end(self._make_result(content=""))
-        assert handler.successful_requests == 1
-        assert handler.empty_responses == 1
-        assert handler.models_token_dict["ollama"]["gpt-oss:20B"]["empty_responses"] == 1
-
-    @pytest.mark.asyncio
-    async def test_on_llm_end_does_not_count_nonempty_response(self):
-        """A response with content is a successful request but not an empty one."""
-        handler = self._make_handler()
-        with owning_agent_scope(handler):
-            await handler.on_llm_end(self._make_result(content="a real answer"))
-        assert handler.successful_requests == 1
-        assert handler.empty_responses == 0
-        assert handler.models_token_dict["ollama"]["gpt-oss:20B"]["empty_responses"] == 0
-
-
-class TestExclusiveModelAttribution:
-    """
-    Test cases for the own-call vs downstream-call attribution in LlmTokenCallbackHandler.
-
-    Handlers are inheritable langchain callbacks, so an agent's handler also receives
-    events for LLM calls made by downstream agents.  Those events must update the
-    scalar subtree totals but NOT models_token_dict, so that merging every agent's
-    models_token_dict into the request-wide accounting counts each call exactly once.
-    """
-
-    CHAT_MODEL_START_SERIALIZED = {"id": ["langchain", "chat_models", "openai", "ChatOpenAI"]}
-
-    def _make_result(self) -> LLMResult:
-        """Build an LLMResult wrapping a single AIMessage with usage metadata."""
-        message = AIMessage(
-            content="an answer",
-            usage_metadata={"input_tokens": 10, "output_tokens": 5, "total_tokens": 15},
-            response_metadata={"model_name": "gpt-4"},
-        )
-        return LLMResult(generations=[[ChatGeneration(message=message)]])
-
-    @pytest.mark.asyncio
-    async def test_own_call_updates_models_and_scalars(self):
-        """An LLM call belonging to this handler's own agent updates both tallies."""
-        handler = LlmTokenCallbackHandler(llm_infos={})
-        with owning_agent_scope(handler):
-            await handler.on_chat_model_start(self.CHAT_MODEL_START_SERIALIZED, [])
-            await handler.on_llm_end(self._make_result())
-
-        assert handler.total_tokens == 15
-        assert handler.successful_requests == 1
-        assert handler.models_token_dict["openai"]["gpt-4"]["total_tokens"] == 15
-        assert handler.models_token_dict["openai"]["gpt-4"]["successful_requests"] == 1
-
-    @pytest.mark.asyncio
-    async def test_downstream_call_updates_scalars_only(self):
-        """
-        A downstream agent's LLM call updates the subtree scalars of an ancestor
-        handler, but not its per-model (exclusive) stats.
-        """
-        ancestor_handler = LlmTokenCallbackHandler(llm_infos={})
-        downstream_handler = LlmTokenCallbackHandler(llm_infos={})
-
-        # At event time, the var holds the handler of the agent that owns the call:
-        # the downstream one.  The ancestor hears the same events through callback
-        # inheritance.
-        with owning_agent_scope(downstream_handler):
-            for handler in (downstream_handler, ancestor_handler):
-                await handler.on_chat_model_start(self.CHAT_MODEL_START_SERIALIZED, [])
-                await handler.on_llm_end(self._make_result())
-
-        # The owning agent counts the call in both tallies.
-        assert downstream_handler.total_tokens == 15
-        assert downstream_handler.models_token_dict["openai"]["gpt-4"]["total_tokens"] == 15
-
-        # The ancestor's subtree totals include the downstream call...
-        assert ancestor_handler.total_tokens == 15
-        assert ancestor_handler.successful_requests == 1
-        # ... but its per-model stats do not: the call is not its own.
-        assert not ancestor_handler.models_token_dict
-        # And its per-call timer state was never touched.
-        assert ancestor_handler.start_time is None
-
-    @pytest.mark.asyncio
-    async def test_call_outside_any_agent_scope_updates_scalars_only(self):
-        """With no owning agent scope at all, treat the call as not our own."""
-        handler = LlmTokenCallbackHandler(llm_infos={})
-
-        await handler.on_chat_model_start(self.CHAT_MODEL_START_SERIALIZED, [])
-        await handler.on_llm_end(self._make_result())
-
-        assert handler.total_tokens == 15
-        assert not handler.models_token_dict
